Jetty9 - Implemented abstract methods on SelectorManager that were always implemented in the same way, making them consistent.

+++ jetty-server/src/main/java/org/eclipse/jetty/server/Connector.java
@@ -4,11 +4,11 @@
 // All rights reserved. This program and the accompanying materials
 // are made available under the terms of the Eclipse Public License v1.0
 // and Apache License v2.0 which accompanies this distribution.
-// The Eclipse Public License is available at 
+// The Eclipse Public License is available at
 // http://www.eclipse.org/legal/epl-v10.html
 // The Apache License v2.0 is available at
 // http://www.opensource.org/licenses/apache2.0.php
-// You may elect to redistribute this code under either of these licenses. 
+// You may elect to redistribute this code under either of these licenses.
 // ========================================================================
 
 package org.eclipse.jetty.server;
@@ -21,47 +21,47 @@ import org.eclipse.jetty.util.component.LifeCycle;
 
 /** HTTP Connector.
  * Implementations of this interface provide connectors for the HTTP protocol.
- * A connector receives requests (normally from a socket) and calls the 
+ * A connector receives requests (normally from a socket) and calls the
  * handle method of the Handler object.  These operations are performed using
  * threads from the ThreadPool set on the connector.
- * 
+ *
  * When a connector is registered with an instance of Server, then the server
  * will set itself as both the ThreadPool and the Handler.  Note that a connector
  * can be used without a Server if a thread pool and handler are directly provided.
- * 
+ *
  */
 public interface Connector extends LifeCycle
-{ 
+{
     /* ------------------------------------------------------------ */
     /**
      * @return the name of the connector. Defaults to the HostName:port
      */
     String getName();
-    
+
     /* ------------------------------------------------------------ */
     Server getServer();
 
     /* ------------------------------------------------------------ */
-    Executor findExecutor();    
-    
+    Executor findExecutor();
+
     /* ------------------------------------------------------------ */
-    Executor getExecutor();    
-    
+    Executor getExecutor();
+
     /* ------------------------------------------------------------ */
-    ByteBufferPool getByteBufferPool();    
-    
+    ByteBufferPool getByteBufferPool();
+
     /* ------------------------------------------------------------ */
     /**
      * @return Max Idle time for connections in milliseconds
      */
-    int getMaxIdleTime();
-    
+    long getMaxIdleTime();
+
     /* ------------------------------------------------------------ */
     /**
      * @return the underlying socket, channel, buffer etc. for the connector.
      */
     Object getTransport();
-    
+
     /* ------------------------------------------------------------ */
     Statistics getStatistics();
 
@@ -69,28 +69,28 @@ public interface Connector extends LifeCycle
     {
         /* ------------------------------------------------------------ */
         /**
-         * Opens the connector 
+         * Opens the connector
          * @throws IOException
          */
         void open() throws IOException;
 
         /* ------------------------------------------------------------ */
         void close();
-        
+
         /* ------------------------------------------------------------ */
         /**
-         * @return The hostname representing the interface to which 
+         * @return The hostname representing the interface to which
          * this connector will bind, or null for all interfaces.
          */
         String getHost();
-        
+
         /* ------------------------------------------------------------ */
         /**
          * @return The configured port for the connector or 0 if any available
          * port may be used.
          */
         int getPort();
-        
+
         /* ------------------------------------------------------------ */
         /**
          * @return The actual port the connector is listening on or
@@ -98,20 +98,20 @@ public interface Connector extends LifeCycle
          */
         int getLocalPort();
     }
-    
+
     interface Statistics extends LifeCycle
     {
         /* ------------------------------------------------------------ */
-        /** 
+        /**
          * @return True if statistics collection is turned on.
          */
         boolean getStatsOn();
-        
+
         /* ------------------------------------------------------------ */
         /** Reset statistics.
          */
         void statsReset();
-        
+
         /* ------------------------------------------------------------ */
         /**
          * @return Get the number of messages received by this connector
@@ -119,7 +119,7 @@ public interface Connector extends LifeCycle
          * is undefined.
          */
         public int getMessagesIn();
-        
+
         /* ------------------------------------------------------------ */
         /**
          * @return Get the number of messages sent by this connector
@@ -127,7 +127,7 @@ public interface Connector extends LifeCycle
          * is undefined.
          */
         public int getMessagesOut();
-        
+
         /* ------------------------------------------------------------ */
         /**
          * @return Get the number of bytes received by this connector
@@ -135,7 +135,7 @@ public interface Connector extends LifeCycle
          * is undefined.
          */
         public int getBytesIn();
-        
+
         /* ------------------------------------------------------------ */
         /**
          * @return Get the number of bytes sent by this connector
@@ -151,42 +151,42 @@ public interface Connector extends LifeCycle
         public long getConnectionsDurationTotal();
 
         /* ------------------------------------------------------------ */
-        /** 
+        /**
          * @return Number of connections accepted by the server since
          * statsReset() called. Undefined if setStatsOn(false).
          */
         public int getConnections() ;
 
         /* ------------------------------------------------------------ */
-        /** 
+        /**
          * @return Number of connections currently open that were opened
          * since statsReset() called. Undefined if setStatsOn(false).
          */
         public int getConnectionsOpen() ;
 
         /* ------------------------------------------------------------ */
-        /** 
+        /**
          * @return Maximum number of connections opened simultaneously
          * since statsReset() called. Undefined if setStatsOn(false).
          */
         public int getConnectionsOpenMax() ;
 
         /* ------------------------------------------------------------ */
-        /** 
+        /**
          * @return Maximum duration in milliseconds of an open connection
          * since statsReset() called. Undefined if setStatsOn(false).
          */
         public long getConnectionsDurationMax();
 
         /* ------------------------------------------------------------ */
-        /** 
+        /**
          * @return Mean duration in milliseconds of open connections
          * since statsReset() called. Undefined if setStatsOn(false).
          */
         public double getConnectionsDurationMean() ;
 
         /* ------------------------------------------------------------ */
-        /** 
+        /**
          * @return Standard deviation of duration in milliseconds of
          * open connections since statsReset() called. Undefined if
          * setStatsOn(false).
@@ -194,28 +194,28 @@ public interface Connector extends LifeCycle
         public double getConnectionsDurationStdDev() ;
 
         /* ------------------------------------------------------------ */
-        /** 
+        /**
          * @return Mean number of messages received per connection
          * since statsReset() called. Undefined if setStatsOn(false).
          */
         public double getConnectionsMessagesInMean() ;
 
         /* ------------------------------------------------------------ */
-        /** 
+        /**
          * @return Standard Deviation of number of messages received per connection
          * since statsReset() called. Undefined if setStatsOn(false).
          */
         public double getConnectionsMessagesInStdDev() ;
 
         /* ------------------------------------------------------------ */
-        /** 
+        /**
          * @return Maximum number of messages received per connection
          * since statsReset() called. Undefined if setStatsOn(false).
          */
         public int getConnectionsMessagesInMax();
-        
+
         /* ------------------------------------------------------------ */
-        /** 
+        /**
          * @return Timestamp stats were started at.
          */
         public long getStatsOnMs();
